+++ rabbithole/src/entity.rs
@@ -1,12 +1,34 @@
-use crate::model::document::{Document, DocumentItem, Included, PrimaryDataItem};
+use crate::model::document::{Document, Included};
 use crate::model::link::{Link, Links};
+use crate::model::query::{FieldsQuery, IncludeQuery};
 use crate::model::relationship::{RelationshipLinks, Relationships};
 use crate::model::resource::{Attributes, Resource, ResourceIdentifier};
-use crate::model::Id;
 use crate::RbhOptionRes;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
+
 use std::collections::{HashMap, HashSet};
 use std::iter::FromIterator;
+
+fn filter_included(
+    included: Included, relationships: &Relationships, included_fields: &IncludeQuery,
+) -> Included {
+    if !included_fields.is_empty() {
+        // Get res_ids from  which are in the `included_fields` array
+        let needed_res_ids: HashSet<ResourceIdentifier> = included_fields
+            .iter()
+            .filter_map(|fid| relationships.get(fid))
+            .flat_map(|r| r.data.data())
+            .collect();
+        included
+            .into_iter()
+            .filter_map(|inc| {
+                needed_res_ids.iter().find_map(|rid| inc.is_resource_id(rid).cloned())
+            })
+            .collect()
+    } else {
+        included
+    }
+}
 
 pub trait Entity: Serialize {
     #[doc(hidden)]
@@ -28,45 +50,18 @@ pub trait Entity: Serialize {
     fn relationships(&self, uri: &str) -> RbhOptionRes<Relationships>;
     #[doc(hidden)]
     fn included(
-        &self, uri: &str, sparse_fields: &HashMap<String, HashSet<String>>,
+        &self, uri: &str, included_fields: &IncludeQuery, sparse_fields: &FieldsQuery,
     ) -> RbhOptionRes<Included>;
 
     fn to_document(
-        &self, uri: &str, included_fields: &HashSet<String>,
-        sparse_fields: &HashMap<String, HashSet<String>>,
+        &self, uri: &str, included_fields: &IncludeQuery, sparse_fields: &FieldsQuery,
     ) -> RbhOptionRes<Document> {
-        if let Some(item) = self.to_document_item(uri, included_fields, sparse_fields)? {
-            Ok(Some(Document { item, ..Default::default() }))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn to_document_item(
-        &self, uri: &str, included_fields: &HashSet<String>,
-        sparse_fields: &HashMap<String, HashSet<String>>,
-    ) -> RbhOptionRes<DocumentItem> {
-        if let (Some(res), Some(included)) =
-            (self.to_resource(uri, sparse_fields)?, self.included(uri, sparse_fields)?)
-        {
-            let included: Included = if !included_fields.is_empty() {
-                let needed_res_ids: HashSet<ResourceIdentifier> = included_fields
-                    .iter()
-                    .filter_map(|fid| res.relationships.get(fid))
-                    .flat_map(|r| r.data.data())
-                    .collect();
-                included
-                    .into_iter()
-                    .filter_map(|inc| {
-                        needed_res_ids.iter().find_map(|rid| inc.is_resource_id(rid).cloned())
-                    })
-                    .collect()
-            } else {
-                included
-            };
-
-            let primary = (PrimaryDataItem::Single(Box::new(res)), included);
-            Ok(Some(DocumentItem::PrimaryData(Some(primary))))
+        if let (Some(res), Some(included)) = (
+            self.to_resource(uri, sparse_fields)?,
+            self.included(uri, included_fields, sparse_fields)?,
+        ) {
+            let included = filter_included(included, &res.relationships, included_fields);
+            Ok(Some(Document::single_resource(res, included)))
         } else {
             Ok(None)
         }
@@ -108,9 +103,7 @@ pub trait Entity: Serialize {
         }
     }
 
-    fn to_resource(
-        &self, uri: &str, sparse_fields: &HashMap<String, HashSet<String>>,
-    ) -> RbhOptionRes<Resource> {
+    fn to_resource(&self, uri: &str, sparse_fields: &FieldsQuery) -> RbhOptionRes<Resource> {
         if let (Some(ty), Some(id), Some(mut attributes), Some(mut relationships), Some(links)) =
             (self.ty(), self.id(), self.attributes(), self.relationships(uri)?, self.links(uri)?)
         {
@@ -146,10 +139,10 @@ impl<T: Entity> Entity for Option<T> {
     }
 
     fn included(
-        &self, uri: &str, sparse_fields: &HashMap<String, HashSet<String>>,
+        &self, uri: &str, included_fields: &IncludeQuery, sparse_fields: &FieldsQuery,
     ) -> RbhOptionRes<Included> {
         if let Some(s) = self.as_ref() {
-            s.included(uri, sparse_fields)
+            s.included(uri, included_fields, sparse_fields)
         } else {
             Ok(None)
         }
@@ -168,16 +161,13 @@ impl<T: Entity> Entity for Box<T> {
     }
 
     fn included(
-        &self, uri: &str, sparse_fields: &HashMap<String, HashSet<String>>,
+        &self, uri: &str, included_fields: &IncludeQuery, sparse_fields: &FieldsQuery,
     ) -> RbhOptionRes<Included> {
-        self.as_ref().included(uri, sparse_fields)
+        self.as_ref().included(uri, included_fields, sparse_fields)
     }
 }
 
-impl<'de, T: Entity> Entity for &T
-where
-    T: Deserialize<'de>,
-{
+impl<T: Entity> Entity for &T {
     fn ty(&self) -> Option<String> { (*self).ty() }
 
     fn id(&self) -> Option<String> { (*self).id() }
@@ -187,8 +177,65 @@ where
     fn relationships(&self, uri: &str) -> RbhOptionRes<Relationships> { (*self).relationships(uri) }
 
     fn included(
-        &self, uri: &str, sparse_fields: &HashMap<String, HashSet<String>>,
+        &self, uri: &str, included_fields: &IncludeQuery, sparse_fields: &FieldsQuery,
     ) -> RbhOptionRes<Included> {
-        (*self).included(uri, sparse_fields)
+        (*self).included(uri, included_fields, sparse_fields)
+    }
+}
+
+impl<T: Entity> Entity for [T] {
+    fn ty(&self) -> Option<String> { None }
+
+    fn id(&self) -> Option<String> { None }
+
+    fn attributes(&self) -> Option<Attributes> { None }
+
+    fn relationships(&self, _: &str) -> RbhOptionRes<Relationships> { Ok(None) }
+
+    fn included(
+        &self, uri: &str, included_fields: &IncludeQuery, sparse_fields: &FieldsQuery,
+    ) -> RbhOptionRes<Included> {
+        let mut hashmap: HashMap<ResourceIdentifier, Resource> = Default::default();
+        for e in self {
+            if let Some(inc) = e.included(uri, included_fields, sparse_fields)? {
+                for i in inc {
+                    hashmap.insert(ResourceIdentifier { ty: i.ty.clone(), id: i.id.clone() }, i);
+                }
+            }
+        }
+
+        Ok(Some(hashmap.values().cloned().collect()))
+    }
+
+    fn to_document(
+        &self, uri: &str, included_fields: &IncludeQuery, sparse_fields: &FieldsQuery,
+    ) -> RbhOptionRes<Document> {
+        if let Some(included) = self.included(uri, included_fields, sparse_fields)? {
+            let mut reses = vec![];
+            for e in self {
+                if let Some(res) = e.to_resource(uri, sparse_fields)? {
+                    reses.push(res);
+                }
+            }
+            Ok(Some(Document::multiple_resources(reses, included)))
+        } else {
+            Ok(None)
+        }
+    }
+}
+
+impl<T: Entity> Entity for Vec<T> {
+    fn ty(&self) -> Option<String> { None }
+
+    fn id(&self) -> Option<String> { None }
+
+    fn attributes(&self) -> Option<Attributes> { None }
+
+    fn relationships(&self, _: &str) -> RbhOptionRes<Relationships> { Ok(None) }
+
+    fn included(
+        &self, uri: &str, included_fields: &IncludeQuery, sparse_fields: &FieldsQuery,
+    ) -> RbhOptionRes<Included> {
+        self.as_slice().included(uri, included_fields, sparse_fields)
     }
 }