+++ rabbithole/src/model/pagination.rs
@@ -1,8 +1,10 @@
+use crate::model::link::RawUri;
+
 /// Pagination links
 #[derive(Serialize, Deserialize, Debug, Eq, PartialEq)]
 pub struct Pagination {
-    pub first: Option<String>,
-    pub prev: Option<String>,
-    pub next: Option<String>,
-    pub last: Option<String>,
+    pub first: Option<RawUri>,
+    pub prev: Option<RawUri>,
+    pub next: Option<RawUri>,
+    pub last: Option<RawUri>,
 }